requirements.txt
flask==2.0.1
flask-alembic==2.0.1
flask-sqlalchemy==2.5.1
prometheus-flask-exporter==0.18.2
pymysql==1.0.2
pyyaml==5.4.1
sqlalchemy==1.4.21
